debug

+++ src/TeamItem.cpp
@@ -140,7 +140,7 @@ size_t TeamItem::CountMemory() {
 
 char *TeamItem::GetFullPath() {
 	image_info info;
-	ssize_t cookie = 0;
+	int32 cookie = 0;
 
 	if (get_next_image_info(team, &cookie, &info) == B_OK) {
 		fprintf(stderr, "team=%d, GetFullPath()=%s\n", (int)team, info.name);
